renamed infra http folder to api
