+++ users-service/src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { APP_GUARD } from '@nestjs/core';
 import { DatabaseModule } from './infra/database/database.module';
-import { HttpModule } from '@infra/http/http.module';
-import { AuthGuard } from '@infra/http/guards/auth.guard';
+import { APIModule } from '@infra/api/api.module';
+import { AuthGuard } from '@infra/api/guards/auth.guard';
 
 @Module({
-  imports: [HttpModule, DatabaseModule],
+  imports: [APIModule, DatabaseModule],
   providers: [
     {
       provide: APP_GUARD,
