+++ users-service/src/infra/api/api.module.ts
@@ -4,4 +4,4 @@ import { UsersModule } from '@users/users.module';
 @Module({
   imports: [UsersModule],
 })
-export class HttpModule {}
+export class APIModule {}
