+++ users-service/src/infra/api/guards/auth.guard.ts
